Migrate common.js to Typescript
